Move to nosetests
update travis script to run nosetests and begin switched examples over
